requirements.txt
# Core Framework
sqlalchemy==2.0.23
flask==3.0.0
streamlit==1.28.0

# AI/NLP
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
protobuf==4.24.4

# HTTP & Environment
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3

# Additional utilities
urllib3==2.0.7
tqdm==4.66.1
